Update the last frame in the first layout
The original purpose of commit c3cf953 is to avoid an
unnecessary resize after activity relaunched. So it
would be more accurate to only update the last frames
when the window is performing the first layout instead
of every relayout. And in this case the position for
the windows in an embedded display should also be updated
as the initialization.

Bug: 144585233
Fix: 145715291
Test: atest DisplaySizeTest
            AccessibilityEmbeddedDisplayTest

Change-Id: Ibf38933099cc73f6f6d3ea8c2afb7070a86a6edc