Return to direct mutation for Provider authority
The EffectiveProvider wrapper is incorrect. The fact that ComponentResolver
mutates the Provider instance is a required behavior, otherwise the authority
doesn't get propagated.

This causes queries to fail because the EffectiveProvider thinks the authority
should be null. This change reverts to the original behavior pre-refactor,
directly mutating the ParsedProvider instance.

Fixing this mutation will have to wait for a followup, unfortunately.

Bug: 146072648

Test: run test in linked bug that queries content providers
Test: manual verified failing case

Change-Id: Ief960bec3692d60e823a60734c2196ee6caeff7a